Return error on stack height limit exceed
